refactor(relay): extract shared event scanner into lib/event-scanner.mjs
