+++ skills/opencode-delegate/scripts/relay.mjs
@@ -71,6 +71,7 @@ import { mkdirSync, writeFileSync, renameSync, readFileSync, existsSync, appendF
 import { join, resolve, basename } from "node:path";
 import { constants, tmpdir } from "node:os";
 import { StringDecoder } from "node:string_decoder";
+import { makeEventScanner } from "../../../lib/event-scanner.mjs";
 
 const VERSION_PROBE_TIMEOUT_MS = 10_000;
 const MAX_TIMER_MS = 2_147_483_647;
@@ -273,64 +274,6 @@ function buildArgv(opts) {
   // No message argument: the brief is piped on stdin (see dispatchToOpenCode),
   // which avoids all argv-quoting issues with multi-line, XML-tagged briefs.
   return argv;
-}
-
-function makeEventScanner(onObject) {
-  // OpenCode emits newline-delimited JSON events on stdout, but local plugins can
-  // prepend terminal-notify escape sequences (e.g. `]777;notify;...{...}`) on the
-  // same line. A plain line-splitter would choke on those. This brace-aware
-  // scanner instead walks the byte stream, ignores anything at depth 0 that isn't
-  // a top-level object, and emits each complete `{...}` it closes — robust to junk
-  // prefixes and concatenated objects alike. String/escape state is tracked so
-  // braces inside string values never throw off the depth count.
-  let buf = "";
-  let depth = 0;
-  let start = -1;
-  let inString = false;
-  let escaped = false;
-  return (chunk) => {
-    buf += chunk;
-    for (let i = 0; i < buf.length; i += 1) {
-      const ch = buf[i];
-      if (inString) {
-        if (escaped) escaped = false;
-        else if (ch === "\\") escaped = true;
-        else if (ch === '"') inString = false;
-        continue;
-      }
-      // Only track strings inside an object (depth > 0). At depth 0 we're skipping a
-      // junk prefix (e.g. a terminal-notify escape), and an unmatched `"` there must
-      // not swallow the real `{...}` that follows in the same chunk.
-      if (ch === '"') { if (depth > 0) inString = true; continue; }
-      if (ch === "{") {
-        if (depth === 0) start = i;
-        depth += 1;
-      } else if (ch === "}") {
-        if (depth > 0) {
-          depth -= 1;
-          if (depth === 0 && start !== -1) {
-            const slice = buf.slice(start, i + 1);
-            try { onObject(JSON.parse(slice)); } catch { /* not a JSON object we care about */ }
-            start = -1;
-          }
-        }
-      }
-    }
-    // Retain only an in-progress object (if any) so the buffer can't grow without
-    // bound; everything already emitted or skipped is dropped. Reset the scanner
-    // state too: the next call re-derives depth/string state by scanning the
-    // retained buffer (which always begins at an object's `{`) from scratch.
-    // Without the reset, the carried-over depth double-counts the retained braces,
-    // so an object split across chunks never closes and its event is lost.
-    // ponytail: O(n^2) if a single object spans many chunks (each re-scans the
-    // retained prefix); fine for OpenCode's event sizes — switch to a suffix-only
-    // scan if it ever bites.
-    buf = depth > 0 && start !== -1 ? buf.slice(start) : "";
-    start = -1;
-    depth = 0;
-    inString = false;
-    escaped = false;
-  };
 }
 
 function prepareRunDir(opts, brief) {
